[FAB-6738] GetState in CouchDB fails to return database
For a while composer has created composite keys in a specific
format and this has worked fine. Since the introduction of a
change to the GetState protocol, GetState has not been returning
information.

Additional investigation showed that the error occurs due to a
bad URL generated when then key contains a colon ":"

Problem is resolved by encoding the path using url EscapedPath()
method instead of the String() method.

Note:  This commit also contains the fix from:
https://gerrit.hyperledger.org/r/#/c/13993 (FAB-6355)

Change-Id: I46a0a85728a36c0a8bd727ad28a75fc02b42206b
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>